Rough draft for OnBoardingPage 1
